+++ jorldy/core/network/_head_dict.txt
@@ -5,6 +5,4 @@ format: (key, class)
 ('cnn_lstm', <class 'core.network.head.CNN_LSTM'>)
 ('mlp', <class 'core.network.head.MLP'>)
 ('mlp_lstm', <class 'core.network.head.MLP_LSTM'>)
-('mlp_residualblock', <class 'core.network.head.MLP_Residualblock'>)
 ('multi', <class 'core.network.head.Multi'>)
-('residualblock', <class 'core.network.head.Residualblock'>)
