+++ jorldy/core/network/_network_dict.txt
@@ -5,6 +5,7 @@ format: (key, class)
 ('continuous_policy_separate_value', <class 'core.network.policy_value.ContinuousPolicySeparateValue'>)
 ('continuous_policy_value', <class 'core.network.policy_value.ContinuousPolicyValue'>)
 ('continuous_q_network', <class 'core.network.q_network.ContinuousQ_Network'>)
+('conv_residualblock', <class 'core.network.muzero.CONV_Residualblock'>)
 ('deterministic_policy', <class 'core.network.policy.DeterministicPolicy'>)
 ('discrete_policy', <class 'core.network.policy.DiscretePolicy'>)
 ('discrete_policy_separate_value', <class 'core.network.policy_value.DiscretePolicySeparateValue'>)
@@ -16,6 +17,7 @@ format: (key, class)
 ('icm_mlp', <class 'core.network.icm.ICM_MLP'>)
 ('icm_multi', <class 'core.network.icm.ICM_Multi'>)
 ('iqn', <class 'core.network.iqn.IQN'>)
+('mlp_residualblock', <class 'core.network.muzero.MLP_Residualblock'>)
 ('muzero_mlp', <class 'core.network.muzero.Muzero_mlp'>)
 ('muzero_resnet', <class 'core.network.muzero.Muzero_Resnet'>)
 ('noisy', <class 'core.network.noisy.Noisy'>)
